Added release.Windows.bat

+++ c/cl.clean.bat
@@ -4,4 +4,4 @@ del *.ilk
 del *.pdb
 del *.exe
 del *.dll
-rd /s /q output
+if exist output rd /s /q output
